meta: update generated files

+++ parser/y.go
@@ -565,7 +565,7 @@ func (l *Lexer) removeEmbeddedStrTokenTypes() {
 }
 
 //line yacctab:1
-var yyExca = [...]int{
+var yyExca = [...]int8{
 	-1, 1,
 	1, -1,
 	-2, 0,
@@ -575,7 +575,7 @@ const yyPrivate = 57344
 
 const yyLast = 2531
 
-var yyAct = [...]int{
+var yyAct = [...]int16{
 	67, 188, 235, 160, 162, 164, 23, 232, 293, 109,
 	166, 115, 83, 165, 2, 121, 122, 82, 54, 4,
 	254, 31, 277, 344, 121, 122, 81, 80, 354, 342,
@@ -832,7 +832,7 @@ var yyAct = [...]int{
 	78,
 }
 
-var yyPact = [...]int{
+var yyPact = [...]int16{
 	896, -1000, -44, 973, -1000, -1000, -63, -1000, 1940, 2249,
 	2249, 2249, 2249, 57, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, 2183, -57, 2249, 2249, 2249, 2249, 2249, -1000,
@@ -871,7 +871,7 @@ var yyPact = [...]int{
 	131, 1940, -1000, -1000, -1000, -1000, -1000, -1000,
 }
 
-var yyPgo = [...]int{
+var yyPgo = [...]int16{
 	0, 330, 13, 19, 329, 328, 326, 278, 324, 318,
 	316, 314, 313, 302, 300, 299, 297, 294, 287, 21,
 	281, 279, 277, 236, 275, 272, 271, 269, 266, 233,
@@ -880,7 +880,7 @@ var yyPgo = [...]int{
 	30, 240, 238, 237, 234, 231, 204,
 }
 
-var yyR1 = [...]int{
+var yyR1 = [...]int8{
 	0, 1, 1, 1, 1, 2, 2, 2, 3, 3,
 	3, 4, 6, 5, 5, 5, 5, 7, 7, 7,
 	7, 7, 16, 16, 16, 16, 16, 16, 42, 42,
@@ -907,7 +907,7 @@ var yyR1 = [...]int{
 	56, 56, 47, 47, 48, 48,
 }
 
-var yyR2 = [...]int{
+var yyR2 = [...]int8{
 	0, 1, 2, 1, 0, 1, 3, 2, 1, 1,
 	1, 1, 3, 2, 2, 2, 2, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 3, 1, 1, 1,
@@ -934,7 +934,7 @@ var yyR2 = [...]int{
 	1, 2, 1, 1, 1, 2,
 }
 
-var yyChk = [...]int{
+var yyChk = [...]int16{
 	-1000, -1, -2, 70, -3, -4, -5, -6, -7, 77,
 	78, 79, 80, -16, -8, -9, -10, -14, -15, -12,
 	-11, -13, 53, -42, 19, 20, 21, 23, 40, -17,
@@ -973,7 +973,7 @@ var yyChk = [...]int{
 	70, -7, 58, 58, 70, 58, 58, 54,
 }
 
-var yyDef = [...]int{
+var yyDef = [...]int16{
 	4, -2, 1, 3, 5, 8, 9, 10, 11, 0,
 	0, 0, 0, 17, 18, 19, 20, 21, 22, 23,
 	24, 25, 0, 27, 0, 0, 0, 0, 0, 32,
@@ -1012,11 +1012,11 @@ var yyDef = [...]int{
 	0, 115, 93, 94, 148, 103, 104, 168,
 }
 
-var yyTok1 = [...]int{
+var yyTok1 = [...]int8{
 	1,
 }
 
-var yyTok2 = [...]int{
+var yyTok2 = [...]int8{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -1028,7 +1028,7 @@ var yyTok2 = [...]int{
 	82, 83, 84, 85, 86,
 }
 
-var yyTok3 = [...]int{
+var yyTok3 = [...]int8{
 	0,
 }
 
@@ -1110,9 +1110,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := yyPact[state]
+	base := int(yyPact[state])
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
+		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -1122,13 +1122,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || yyExca[i+1] != state {
+		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := yyExca[i]
+			tok := int(yyExca[i])
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -1159,30 +1159,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = yyTok1[0]
+		token = int(yyTok1[0])
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = yyTok1[char]
+		token = int(yyTok1[char])
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = yyTok2[char-yyPrivate]
+			token = int(yyTok2[char-yyPrivate])
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = yyTok3[i+0]
+		token = int(yyTok3[i+0])
 		if token == char {
-			token = yyTok3[i+1]
+			token = int(yyTok3[i+1])
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = yyTok2[1] /* unknown char */
+		token = int(yyTok2[1]) /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -1237,7 +1237,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = yyPact[yystate]
+	yyn = int(yyPact[yystate])
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -1248,8 +1248,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = yyAct[yyn]
-	if yyChk[yyn] == yytoken { /* valid shift */
+	yyn = int(yyAct[yyn])
+	if int(yyChk[yyn]) == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -1262,7 +1262,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = yyDef[yystate]
+	yyn = int(yyDef[yystate])
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -1271,18 +1271,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
+			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = yyExca[xi+0]
+			yyn = int(yyExca[xi+0])
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = yyExca[xi+1]
+		yyn = int(yyExca[xi+1])
 		if yyn < 0 {
 			goto ret0
 		}
@@ -1304,10 +1304,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = yyPact[yyS[yyp].yys] + yyErrCode
+				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = yyAct[yyn] /* simulate a shift of "error" */
-					if yyChk[yystate] == yyErrCode {
+					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
+					if int(yyChk[yystate]) == yyErrCode {
 						goto yystack
 					}
 				}
@@ -1343,7 +1343,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= yyR2[yyn]
+	yyp -= int(yyR2[yyn])
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -1354,16 +1354,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = yyR1[yyn]
-	yyg := yyPgo[yyn]
+	yyn = int(yyR1[yyn])
+	yyg := int(yyPgo[yyn])
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = yyAct[yyg]
+		yystate = int(yyAct[yyg])
 	} else {
-		yystate = yyAct[yyj]
-		if yyChk[yystate] != -yyn {
-			yystate = yyAct[yyg]
+		yystate = int(yyAct[yyj])
+		if int(yyChk[yystate]) != -yyn {
+			yystate = int(yyAct[yyg])
 		}
 	}
 	// dummy call; replaced with literal code
